+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Implement `project[X|Y|Z]rand!` as a simpler interface to `project!` with automatic randomization of measurement phases.
 - Implement `sMX`/`sMY`/`sMZ` symbolic measurement operations that can be used with `apply!`. Use `projectrand!` internally.
 - Move `Register` out of `Experimental`. Used with `sMX`/etc to store classical bit results during circuit evolution.
+- **(breaking)** The experimental module `NoisyCircuits` now supports only `MixedDestabilizer` and `Register`.
 
 ## v0.4.3